fix(content): harden material migration
